renamed temp reference function & added test
